2.3.1 v1

+++ src/main/java/web/service/UserService.java
@@ -9,7 +9,7 @@ public interface UserService {
 
     void saveUser(User user);
 
-    User getUser(int id);
+    User getUser(Long id);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 }
